requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Deprecated==1.2.13
Django==4.0.5
django-crispy-forms==1.14.0
django-filter==22.1
django-oauth-toolkit==2.1.0
django-taggit==3.0.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
jwcrypto==1.3.1
Markdown==3.3.7
mysqlclient==2.1.0
oauthlib==3.2.0
pycparser==2.21
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.0
